refactor: ログ情報の読み込み処理を分離
